requirements.txt
# Telegram Bot
aiogram==3.2.0
python-telegram-bot==20.7

# Web scraping and parsing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
selenium==4.15.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3

# Data processing
pandas==2.1.3
numpy==1.25.2

# Text processing and NLP
spacy==3.7.2
natasha==1.8.0
pymorphy2==0.9.1

# Email validation
email-validator==2.1.0
dnspython==2.4.2

# Async and queues
asyncio
aiohttp==3.9.1
# redis==5.0.1  # Опционально, для Redis storage
# celery==5.3.4  # Опционально, для распределенных задач

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
